refactor: logout redirect

+++ src/athena/app/routers/user_router.py
@@ -11,6 +11,7 @@ from constants.mapped_prefix import MAPPED_API_ENDPOINT_PREFIX
 from extensions.base_requests import async_fetch, MethodEnum
 from extensions.env_var import get_env_var
 
+from serializers.auth_serializer import TokenRequestHeaders
 from serializers.context_serializer import AlertTypeEnum
 
 from .base_router import async_render_template
@@ -26,9 +27,7 @@ router = APIRouter(
 @router.get('')
 @requires('authenticated')
 async def user_page(request: Request):
-    return await async_render_template(
-        'user_template.html', request, 200
-    )
+    return await async_render_template('user_template.html', request)
 
 
 @router.get('/login', response_class=HTMLResponse)
@@ -39,9 +38,7 @@ async def user_login_page(request: Request):
             headers={'X-Athena-Flash-Message-Authenticated':
                      'User already been logged!'}
         )
-    return await async_render_template(
-        'login_template.html', request, 200
-    )
+    return await async_render_template('login_template.html', request)
 
 
 @router.post('/login', response_class=RedirectResponse)
@@ -50,16 +47,16 @@ async def user_login(
     response: Response,
     form_data: LoginForm = Depends(LoginForm.as_form)
 ):
-    url = get_env_var("ATHENA_API_BASE_URL", raise_exception=True)
+    base_url = get_env_var('ATHENA_API_BASE_URL', raise_exception=True)
     try:
         data = await async_fetch(
             MethodEnum.post,
-            f'{url}/auth/token',
+            f'{base_url}/auth/token',
             headers={
                 'accept': 'application/json',
                 'Content-Type': 'application/x-www-form-urlencoded'
             },
-            data=form_data.dict(by_alias=True)
+            data=form_data
         )
     except HTTPStatusError:
         return await async_render_template(
@@ -72,7 +69,7 @@ async def user_login(
             'Unexpected error', AlertTypeEnum.danger, exc
         )
     response.set_cookie(
-        key="access_token",
+        key='access_token',
         value=f'{data["token_type"]} {data["access_token"]}',
         httponly=True
     )
@@ -82,10 +79,38 @@ async def user_login(
 
 
 @router.get('/logout')
-@requires('authenticated')
 async def user_logout(request: Request):
+    if not request.user.is_authenticated:
+        return RedirectResponse(
+            request.base_url, 302,
+            headers={'X-Athena-Flash-Message-Not-Authenticated':
+                     'User must be authenticated!'}
+        )
     response = RedirectResponse(
         router.url_path_for('user_login'), 302
     )
     response.delete_cookie(key='access_token')
     return response
+
+
+@router.get('/admin')
+@requires('admin')
+async def user_admin(request: Request):
+    base_url = get_env_var('ATHENA_API_BASE_URL')
+    try:
+        response = await async_fetch(
+            MethodEnum.get, f'{base_url}/admin/admin',
+            headers=TokenRequestHeaders(
+                access_token=request.headers['cookie']
+            )
+        )
+    except HTTPStatusError:
+        return await async_render_template(
+            'errors/403_error.html', request, 403,
+            'Not authorized!', AlertTypeEnum.danger
+        )
+
+    return await async_render_template(
+        'admin_template.html', request,
+        context_request={'response': response}
+    )
